+++ R/create_report.r
@@ -57,6 +57,7 @@
 #' # Load library
 #' library(ggplot2)
 #' library(data.table)
+#' library(rmarkdown)
 #'
 #' # Set some missing values
 #' diamonds2 <- data.table(diamonds)